Create CatAndAMouse.java
Two cats named A and B are standing at integral points on the x-axis. Cat A is standing at point x and cat B is standing at point y. Both cats run at the same speed, and they want to catch a mouse named C that's hiding at an integral point z on the x-axis. Can you determine who will catch the mouse?

You are given q queries in the form of x,y, and z. For each query, print the appropriate answer on a new line:

If cat A catches the mouse first, print Cat A.
If cat B catches the mouse first, print Cat B.
If both cats reach the mouse at the same time, print Mouse C as the two cats fight and mouse escapes.